rename /query/type.ts to types.ts
